добавил интерфейсы и имплементации для сервисов
