+++ src/main/java/org/example/controller/SubscriptionController.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.example.controlleradvice.CommonErrorApiResponses;
 import org.example.dto.request.RequestSubscriptionDTO;
 import org.example.dto.response.ResponseSubscriptionDTO;
-import org.example.service.SubscriptionService;
+import org.example.service.impl.SubscriptionServiceImpl;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -20,7 +20,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class SubscriptionController {
 
-    private final SubscriptionService subscriptionService;
+    private final SubscriptionServiceImpl subscriptionServiceImpl;
 
     @Operation(summary = "Получение топ 3 подписок по кол-ву пользователей")
     @ApiResponse(responseCode = "200", description = "Успешное выполнение операции",
@@ -29,7 +29,7 @@ public class SubscriptionController {
     @CommonErrorApiResponses
     @GetMapping("/top")
     public ResponseEntity<List<ResponseSubscriptionDTO>> getTopOfSubscriptions() {
-        return ResponseEntity.ok(subscriptionService.getTopOfSubscriptions());
+        return ResponseEntity.ok(subscriptionServiceImpl.getTopOfSubscriptions());
     }
 
 
@@ -40,6 +40,6 @@ public class SubscriptionController {
     @CommonErrorApiResponses
     @PostMapping
     public ResponseEntity<ResponseSubscriptionDTO> createSubscription(@RequestBody RequestSubscriptionDTO requestSubscriptionDTO) {
-        return ResponseEntity.ok(subscriptionService.createSubscription(requestSubscriptionDTO));
+        return ResponseEntity.ok(subscriptionServiceImpl.createSubscription(requestSubscriptionDTO));
     }
 }
