+++ src/main/java/org/example/controller/UserController.java
@@ -9,7 +9,7 @@ import org.example.controlleradvice.CommonErrorApiResponses;
 import org.example.controlleradvice.CommonErrorApiResponsesWith404;
 import org.example.dto.request.RequestUserDTO;
 import org.example.dto.response.ResponseUserDTO;
-import org.example.service.UserService;
+import org.example.service.impl.UserServiceImpl;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class UserController {
 
-    private final UserService userService;
+    private final UserServiceImpl userServiceImpl;
 
     @Operation(summary = "Получение пользователя по id")
     @ApiResponse(responseCode = "200", description = "Успешное выполнение операции",
@@ -28,7 +28,7 @@ public class UserController {
     @CommonErrorApiResponsesWith404
     @GetMapping("/{id}")
     public ResponseEntity<ResponseUserDTO> getUserById(@PathVariable("id") Long id) {
-        return ResponseEntity.ok(userService.getUserById(id));
+        return ResponseEntity.ok(userServiceImpl.getUserById(id));
     }
 
     @Operation(summary = "Создание пользователя")
@@ -38,7 +38,7 @@ public class UserController {
     @CommonErrorApiResponses
     @PostMapping
     public ResponseEntity<ResponseUserDTO> createUser(@RequestBody RequestUserDTO userDTO) {
-        return ResponseEntity.ok(userService.createUser(userDTO));
+        return ResponseEntity.ok(userServiceImpl.createUser(userDTO));
     }
 
     @Operation(summary = "Обновление пользователя по id")
@@ -48,7 +48,7 @@ public class UserController {
     @CommonErrorApiResponsesWith404
     @PutMapping("/{id}")
     public ResponseEntity<ResponseUserDTO> updateUser(@PathVariable Long id, @RequestBody RequestUserDTO userDTO) {
-        return ResponseEntity.ok(userService.updateUser(id, userDTO));
+        return ResponseEntity.ok(userServiceImpl.updateUser(id, userDTO));
     }
 
     @Operation(summary = "Удаление пользователя по id")
@@ -56,6 +56,6 @@ public class UserController {
     @CommonErrorApiResponsesWith404
     @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable Long id) {
-        userService.deleteUser(id);
+        userServiceImpl.deleteUser(id);
     }
 }
